[SID:3677] fix: InlineSaveIndicator를 형제 모듈로 분리(page.tsx named export 금지)
카디르 CI 실물(head 3e5b954d) — app-router page.tsx는 default export+정해진
config export만 허용돼 `export function InlineSaveIndicator`(테스트 격리용
named export)가 Next build type check(「Page does not match the required
types of a Next.js Page」)를 깼다.

처방: 컴포넌트를 형제 모듈 inline-save-indicator.tsx로 옮기고 page.tsx는
import만(동작 무변 — 코드 diff 0, 파일 이동뿐). 테스트도 그 모듈에서 import.
page.tsx named export 0 확認(grep). `next typegen`+tsc·eslint·vitest 재확認
클린. next build 자체는 로컬 환경(firebase 모듈 해석 pnpm 이슈)이 무관 사유로
막혀 typegen+tsc(.next/types/validator.ts 포함)로 대체 검증 — CI가 최종 판정.
